SCI-6158: extract anonymous classes into standalone files
