+++ org.dawnsci.datavis.model/src/org/dawnsci/datavis/model/PlotModeDataTable2D.java
@@ -27,81 +27,7 @@ public class PlotModeDataTable2D extends PlotModeImage {
 		IDataset d = data[0];
 		
 		
-		ITableDataTrace trace = new ITableDataTrace() {
-			
-			private Object userObject = null;
-			
-			@Override
-			public void setName(String name) {
-				
-			}
-			
-			@Override
-			public String getName() {
-				return null;
-			}
-			
-			@Override
-			public void setVisible(boolean isVisible) {
-				
-			}
-			
-			@Override
-			public void setUserTrace(boolean isUserTrace) {
-				
-			}
-			
-			@Override
-			public void setUserObject(Object userObject) {
-				this.userObject = userObject;
-				
-			}
-			
-			@Override
-			public void setDataName(String name) {
-				
-			}
-			
-			@Override
-			public boolean isVisible() {
-				return false;
-			}
-			
-			@Override
-			public boolean isUserTrace() {
-				return false;
-			}
-			
-			@Override
-			public boolean is3DTrace() {
-				return false;
-			}
-			
-			@Override
-			public Object getUserObject() {
-				return userObject;
-			}
-			
-			@Override
-			public int getRank() {
-				return 0;
-			}
-			
-			@Override
-			public String getDataName() {
-				return null;
-			}
-			
-			@Override
-			public IDataset getData() {
-				return d;
-			}
-			
-			@Override
-			public void dispose() {
-				
-			}
-		};
+		ITableDataTrace trace = new TableDataTrace2D(d);
 
 		trace.setUserObject(userObject);
 		system.addTrace(trace);
